rewrite runintransaction due some issues found in nested, switch, etc

+++ kit_test.go
@@ -40,7 +40,11 @@ func TestRunInTransaction(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
 		func(ctx context.Context) (context.Context, error) {
@@ -61,6 +65,73 @@ func TestRunInTransaction(t *testing.T) {
 		t.Fail()
 	}
 }
+
+func TestConsecutiveRunInTransaction(t *testing.T) {
+	ctx := context.WithValue(context.Background(),
+		gosl.SQL_KEY,
+		gosl.NewQueryable(gosl.ConnectToDB(
+			"root",
+			"abcd",
+			"localhost",
+			"3306",
+			"test_1",
+			1,
+			1,
+			2*time.Minute,
+			2*time.Minute,
+		)))
+	kit := gosl.New(ctx)
+	queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+	_, err := queryable.ExecContext(ctx, "DELETE FROM `hello`")
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	_, err = queryable.ExecContext(ctx, "DELETE FROM `hello_2`")
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	err = kit.RunInTransaction(
+		ctx,
+		func(ctx context.Context) (context.Context, error) {
+			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tiga')")
+			if err != nil {
+				return ctx, err
+			}
+			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empat')")
+			if err != nil {
+				return ctx, err
+			}
+			return ctx, nil
+		},
+	)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+	err = kit.RunInTransaction(
+		ctx,
+		func(ctx context.Context) (context.Context, error) {
+			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('tiga')")
+			if err != nil {
+				return ctx, err
+			}
+			_, err = queryable.ExecContext(ctx, "INSERT INTO `hello_2` VALUES('empat')")
+			if err != nil {
+				return ctx, err
+			}
+			return ctx, nil
+		},
+	)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
+}
+
 func TestRunInTransactionWithSwitchContext(t *testing.T) {
 	ctx := context.WithValue(context.Background(),
 		gosl.SQL_KEY,
@@ -95,33 +166,49 @@ func TestRunInTransactionWithSwitchContext(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	queryable = kit.ContextSwitch(ctx, TKey).Value(gosl.SQL_KEY).(*gosl.Queryable)
-	_, err = queryable.ExecContext(ctx, "DELETE FROM `world`")
-	if err != nil {
-		log.Fatal(err.Error())
-		t.Fail()
-	}
-	ctx = kit.ContextReset(ctx)
-	err = kit.RunInTransaction(
-		ctx,
-		func(ctx context.Context) (context.Context, error) {
-			queryable := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
-			_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('sepuluh')")
+	if ctx, err := kit.ContextSwitch(ctx, TKey); err == nil {
+		if queryable, ok := ctx.Value(gosl.SQL_KEY).(*gosl.Queryable); ok {
+			_, err = queryable.ExecContext(ctx, "DELETE FROM `world`")
 			if err != nil {
-				return ctx, err
+				log.Fatal(err.Error())
+				t.Fail()
 			}
-			queryable = kit.ContextSwitch(ctx, TKey).Value(gosl.SQL_KEY).(*gosl.Queryable)
-			_, err = queryable.ExecContext(ctx, "INSERT INTO `world` VALUES('empat')")
+			ctx, err = kit.ContextReset(ctx)
 			if err != nil {
-				return ctx, err
+				log.Fatal(err.Error())
+				t.Fail()
 			}
-			return ctx, nil
-		},
-	)
-	if err != nil {
-		log.Fatal(err.Error())
-		t.Fail()
+			err = kit.RunInTransaction(
+				ctx,
+				func(ctx context.Context) (context.Context, error) {
+					queryable, ok = ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+					if !ok {
+						t.Fatal("failed to get queryable")
+					}
+					_, err := queryable.ExecContext(ctx, "INSERT INTO `hello` VALUES('sepuluh')")
+					if err != nil {
+						return ctx, err
+					}
+					if ctx, err = kit.ContextSwitch(ctx, TKey); err == nil {
+						queryable, ok = ctx.Value(gosl.SQL_KEY).(*gosl.Queryable)
+						if !ok {
+							t.Fatal("failed to get queryable")
+						}
+					}
+					_, err = queryable.ExecContext(ctx, "INSERT INTO `world` VALUES('empat')")
+					if err != nil {
+						return ctx, err
+					}
+					return ctx, nil
+				},
+			)
+			if err != nil {
+				log.Fatal(err.Error())
+				t.Fail()
+			}
+		}
 	}
+
 }
 
 func TestNestedRunInTransaction(t *testing.T) {
@@ -150,7 +237,11 @@ func TestNestedRunInTransaction(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
 		func(ctx context.Context) (context.Context, error) {
@@ -208,7 +299,11 @@ func TestNestedRunInTransactionWithFailAtTheEnd(t *testing.T) {
 		log.Fatal(err.Error())
 		t.Fail()
 	}
-	ctx = kit.ContextReset(ctx)
+	ctx, err = kit.ContextReset(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+		t.Fail()
+	}
 	err = kit.RunInTransaction(
 		ctx,
 		func(ctx context.Context) (context.Context, error) {
